Fixed legend being unreadable on small plots
